move test inside __tests__ directory
