+++ pages/_app.js
@@ -12,7 +12,7 @@ function MyApp({ Component, pageProps }) {
         <>
             <Head>
                 <link rel="dns-prefetch" href="//fonts.googleapis.com" />
-                <link rel="preconnect" href="https://fonts.gstatic.com/" crossOrigin />
+                <link rel="preconnect" href="https://fonts.gstatic.com/" crossOrigin="true" />
                 <link
                     href="https://fonts.googleapis.com/css2?family=Lato:wght@400;700&display=swap"
                     rel="stylesheet"
